refactor : regacy code remove
