+++ src/main/java/kitchenpos/order/dto/OrderRequest.java
@@ -14,19 +14,6 @@ public class OrderRequest {
         this.orderLineItemRequests = orderLineItemRequests;
     }
 
-    /*public List<Long> getMenuIds() {
-        if(CollectionUtils.isEmpty(orderLineItemRequests)){
-            throw new EmptyOrderLineItemsException();
-        }
-        return orderLineItemRequests.stream()
-                .map(OrderLineItemRequest::getMenuId)
-                .collect(Collectors.toList());
-    }
-
-    public boolean isEqualsMenuSize(long countByIdIn) {
-        return orderLineItemRequests.size() == countByIdIn;
-    }*/
-
     public Long getOrderTableId() {
         return orderTableId;
     }
