+++ src/main/java/kitchenpos/order/util/OrderMapper.java
@@ -5,7 +5,6 @@ import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderLineItems;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
-import kitchenpos.order.exception.EmptyOrderLineItemsException;
 
 import java.util.List;
 import java.util.stream.Collectors;
