+++ src/main/java/kitchenpos/table/domain/OrderTableRepository.java
@@ -3,10 +3,7 @@ package kitchenpos.table.domain;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
     List<OrderTable> findAllByIdIn(List<Long> orderTableIds);
-
-    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
 }
