+++ src/main/java/kitchenpos/table/dto/TableGroupRequest.java
@@ -1,9 +1,6 @@
 package kitchenpos.table.dto;
 
-import org.springframework.util.CollectionUtils;
-
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class TableGroupRequest {
     private List<OrderTableRequest> orderTableRequests;
@@ -13,15 +10,6 @@ public class TableGroupRequest {
 
     public TableGroupRequest(List<OrderTableRequest> orderTableRequests) {
         this.orderTableRequests = orderTableRequests;
-    }
-
-    public List<Long> getOrderTableIds() {
-        if (CollectionUtils.isEmpty(orderTableRequests) || orderTableRequests.size() < 2) {
-            throw new IllegalArgumentException("주문 테이블은 2개 이상이어야 합니다.");
-        }
-        return orderTableRequests.stream()
-                .map(OrderTableRequest::getId)
-                .collect(Collectors.toList());
     }
 
     public boolean isSameOrderTableCount(int orderTableSize) {
