+++ src/main/java/kitchenpos/table/exception/NotChangeToEmptyThatGroupTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.exception;
+package kitchenpos.table.exception;
 
 public class NotChangeToEmptyThatGroupTable extends RuntimeException {
     public NotChangeToEmptyThatGroupTable() {
