+++ src/test/java/kitchenpos/test/application/OrderTableServiceTest.java
@@ -4,7 +4,7 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderLineItems;
 import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.exception.NotChangeToEmptyThatGroupTable;
+import kitchenpos.table.exception.NotChangeToEmptyThatGroupTable;
 import kitchenpos.table.application.OrderTableService;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
